Remove --slim flag from uninstall command

+++ pkg/standalone/uninstall.go
@@ -9,21 +9,23 @@ import (
 	"github.com/dapr/cli/utils"
 )
 
-func removeContainers(uninstallAll bool, dockerNetwork string) []error {
+func removeContainers(uninstallPlacementContainer, uninstallAll bool, dockerNetwork string) []error {
 	var containerErrs []error
 	var err error
 
-	containerErrs = removeDockerContainer(containerErrs, DaprPlacementContainerName, dockerNetwork)
+	if uninstallPlacementContainer {
+		containerErrs = removeDockerContainer(containerErrs, DaprPlacementContainerName, dockerNetwork)
 
-	_, err = utils.RunCmdAndWait(
-		"docker", "rmi",
-		"--force",
-		daprDockerImageName)
+		_, err = utils.RunCmdAndWait(
+			"docker", "rmi",
+			"--force",
+			daprDockerImageName)
 
-	if err != nil {
-		containerErrs = append(
-			containerErrs,
-			fmt.Errorf("could not remove %s image: %s", daprDockerImageName, err))
+		if err != nil {
+			containerErrs = append(
+				containerErrs,
+				fmt.Errorf("could not remove %s image: %s", daprDockerImageName, err))
+		}
 	}
 
 	if uninstallAll {
@@ -35,6 +37,11 @@ func removeContainers(uninstallAll bool, dockerNetwork string) []error {
 }
 
 func removeDockerContainer(containerErrs []error, containerName, network string) []error {
+	exists, _ := confirmContainerIsRunningOrExists(containerName, false)
+	if !exists {
+		fmt.Printf("WARNING: %s container does not exist\n", containerName)
+		return containerErrs
+	}
 	fmt.Println("removing container: ", containerName)
 	_, err := utils.RunCmdAndWait(
 		"docker", "rm",
@@ -74,27 +81,28 @@ func removeInstalledBinaries(binaryFilePrefix, installLocation string) (string, 
 
 // Uninstall reverts all changes made by init. Deletes all installed containers, removes default dapr folder,
 // removes the installed binary and unsets env variables.
-func Uninstall(uninstallSlim, uninstallAll bool, installLocation, dockerNetwork string) error {
+func Uninstall(uninstallAll bool, installLocation, dockerNetwork string) error {
 	var containerErrs []error
 	var err error
 	var path string
-
-	dockerInstalled := false
-	if !uninstallSlim {
-		dockerInstalled = utils.IsDockerInstalled()
-		if dockerInstalled {
-			containerErrs = removeContainers(uninstallAll, dockerNetwork)
-		}
-	}
 
 	path, err = removeInstalledBinaries(daprRuntimeFilePrefix, installLocation)
 	if err != nil {
 		fmt.Println("WARNING: could not delete binary file: ", path)
 	}
 
+	placementFilePath := binaryFilePath(placementServiceFilePrefix, installLocation)
+	_, placementErr := os.Stat(placementFilePath) // check if the placement binary exists
+	uninstallPlacementContainer := os.IsNotExist(placementErr)
 	path, err = removeInstalledBinaries(placementServiceFilePrefix, installLocation)
 	if err != nil {
 		fmt.Println("WARNING: could not delete binary file: ", path)
+	}
+
+	dockerInstalled := false
+	dockerInstalled = utils.IsDockerInstalled()
+	if dockerInstalled {
+		containerErrs = removeContainers(uninstallPlacementContainer, uninstallAll, dockerNetwork)
 	}
 
 	err = rundata.DeleteRunDataFile()
@@ -108,8 +116,9 @@ func Uninstall(uninstallSlim, uninstallAll bool, installLocation, dockerNetwork 
 	}
 
 	err = errors.New("uninstall failed")
-	if !uninstallSlim && !dockerInstalled {
-		return fmt.Errorf("%w \n could not connect to Docker. Docker may not be installed or running", err)
+	if uninstallPlacementContainer && !dockerInstalled {
+		// if placement binary did not exist before trying to delete it and not able to connect to docker.
+		return fmt.Errorf("%w \ncould not delete placement service. Either the placement binary is not found, or Docker may not be installed or running", err)
 	}
 
 	if len(containerErrs) == 0 {
